Rename RuuviTag_RAWv1 decoder file
