+++ src/devices/RuuviTag_RAWv1_json.h
